feat(iop): added wrapper for list component and fixed issues
